heroku_files/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu
fastai==2.2.5
streamlit==0.75.0
Pillow==8.0.1
